requirements.txt
anthropic>=0.100.0
openai>=1.0.0
python-telegram-bot[job-queue]>=22.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
openpyxl>=3.1.0
